further streamlining
